requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.9.2
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
anyio=4.7.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=py39h7f8727e_0
arrow=1.3.0=pypi_0
asttokens=3.0.0=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=py39h06a4308_0
attrs=24.2.0=py39h06a4308_0
babel=2.16.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.3=py39h06a4308_0
blas=1.0=mkl
bleach=6.2.0=py39h06a4308_0
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.11.26=h06a4308_0
certifi=2024.8.30=py39h06a4308_0
cffi=1.17.1=py39h1fdaa30_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
comm=0.2.2=pypi_0
contourpy=1.3.0=pypi_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.6.77=0
cuda-runtime=12.1.0=0
cuda-version=12.6=3
cycler=0.12.1=pypi_0
cyrus-sasl=2.1.28=h52b45da_1
dbus=1.13.18=hb2f20db_0
debugpy=1.8.9=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
exceptiongroup=1.2.2=pypi_0
executing=2.1.0=pypi_0
expat=2.6.4=h6a678d5_0
fastjsonschema=2.21.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py39h06a4308_0
flatbuffers=24.3.25=pypi_0
fontconfig=2.14.1=h55d465d_3
fonttools=4.55.3=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
gast=0.6.0=pypi_0
giflib=5.2.2=h5eee18b_0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-pasta=0.2.0=pypi_0
grpcio=1.68.1=pypi_0
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
h11=0.14.0=py39h06a4308_0
h5py=3.12.1=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=py39h06a4308_0
importlib-metadata=8.5.0=py39h06a4308_0
importlib-resources=6.4.5=pypi_0
importlib_metadata=8.5.0=hd3eb1b0_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.5=py39h06a4308_0
ipython=8.18.1=pypi_0
ipywidgets=8.1.5=py39h06a4308_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=py39h06a4308_0
jinja2=3.1.4=py39h06a4308_1
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_3
json5=0.10.0=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=py39h06a4308_0
jsonschema-specifications=2024.10.1=pypi_0
jupyter=1.1.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyter_client=8.6.0=py39h06a4308_0
jupyter_console=6.6.3=py39h06a4308_0
jupyter_core=5.7.2=py39h06a4308_0
jupyter_events=0.10.0=py39h06a4308_0
jupyter_server=2.14.1=py39h06a4308_0
jupyter_server_terminals=0.4.4=py39h06a4308_1
jupyterlab=4.3.3=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.27.3=py39h06a4308_0
jupyterlab_widgets=3.0.13=py39h06a4308_0
keras=3.7.0=pypi_0
kiwisolver=1.4.7=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=3.0=h295c915_0
libabseil=20240116.2=cxx17_h6a678d5_0
libclang=18.1.1=pypi_0
libclang13=14.0.6=default_he11475f_1
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.11.1.6=0
libcups=2.4.2=h2d74bed_1
libcurand=10.3.7.77=0
libcurl=8.9.1=h251f7ec_0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libllvm14=14.0.6=hecde1de_4
libnghttp2=1.57.0=h2d74bed_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libpq=17.2=hdbd6064_0
libprotobuf=4.25.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h097e994_2
libxml2=2.13.5=hfdd30dd_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.7=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=py39h5eee18b_0
matplotlib=3.9.3=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
ml-dtypes=0.4.1=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
munch=4.0.0=pypi_0
mysql=8.4.0=h29a9f33_1
namex=0.0.8=pypi_0
nbclient=0.10.1=pypi_0
nbconvert=7.16.4=py39h06a4308_0
nbformat=5.10.4=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.2.1=py39h06a4308_0
notebook=7.3.1=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=2.0.1=py39h5f9d8c6_1
numpy-base=2.0.1=py39hb5e798b_1
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openldap=2.6.4=h42fbc30_0
openssl=3.0.15=h5eee18b_0
opt-einsum=3.4.0=pypi_0
optree=0.13.1=pypi_0
overrides=7.7.0=pypi_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=py39h06a4308_0
pcre2=10.42=hebb0a14_1
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=11.0.0=py39hfdbf927_0
pip=24.2=py39h06a4308_0
platformdirs=4.3.6=pypi_0
ply=3.11=py39h06a4308_0
prometheus-client=0.21.1=pypi_0
prometheus_client=0.21.0=py39h06a4308_0
prompt-toolkit=3.0.48=pypi_0
prompt_toolkit=3.0.43=hd3eb1b0_0
protobuf=5.29.1=pypi_0
psutil=6.1.0=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure-eval=0.2.3=pypi_0
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.22=pypi_0
pygments=2.18.0=pypi_0
pyparsing=3.2.0=pypi_0
pyqt=5.15.10=py39h6a678d5_0
pyqt5-sip=12.13.0=py39h5eee18b_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.21=he870216_1
python-dateutil=2.9.0post0=py39h06a4308_2
python-fastjsonschema=2.20.0=py39h06a4308_0
python-json-logger=3.2.0=pypi_0
pytorch=2.5.1=py3.9_cuda12.1_cudnn9.1.0_0
pytorch-cuda=12.1=ha16c6d3_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py39h06a4308_0
pyyaml=6.0.2=py39h5eee18b_0
pyzmq=26.2.0=pypi_0
qt-main=5.15.2=hb6262e9_11
qtconsole=5.6.0=py39h06a4308_0
qtpy=2.4.1=py39h06a4308_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.3=py39h06a4308_1
rfc3339-validator=0.1.4=py39h06a4308_0
rfc3986-validator=0.1.1=py39h06a4308_0
rich=13.9.4=pypi_0
rpds-py=0.22.3=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.13.1=pypi_0
send2trash=1.8.3=pypi_0
setuptools=75.1.0=py39h06a4308_0
sip=6.7.12=py39h6a678d5_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.13.2=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.18.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboard-logger=0.1.0=pypi_0
tensorflow=2.18.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
termcolor=2.5.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.4.0=pypi_0
tk=8.6.14=h39e8969_0
tomli=2.2.1=pypi_0
torchaudio=2.5.1=py39_cu121
torchtriton=3.1.0=py39
torchvision=0.20.1=py39_cu121
tornado=6.4.2=py39h5eee18b_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=py39h06a4308_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.11.0=py39h06a4308_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024.2=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=py39h06a4308_0
visdom=0.2.4=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=py39h06a4308_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py39h06a4308_0
widgetsnbextension=4.0.13=py39h06a4308_0
wrapt=1.17.0=pypi_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.5=h6a678d5_0
zipp=3.21.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
